+++ frontend/pages/configurations/new.vue
@@ -71,7 +71,7 @@ import { Component, Vue } from 'nuxt-property-decorator'
 
 import { mapActions } from 'vuex'
 
-import { SetTitleAction, SetTabsAction } from '@/store/appbar'
+import { SetTitleAction, SetTabsAction, SetShowBackButtonAction } from '@/store/appbar'
 
 import { Configuration } from '@/models/Configuration'
 
@@ -91,7 +91,7 @@ import NonModelOptionsForm, { NonModelOptions } from '@/components/shared/NonMod
   middleware: ['auth'],
   methods: {
     ...mapActions('configurations', ['saveConfiguration', 'createPid']),
-    ...mapActions('appbar', ['setTitle', 'setTabs']),
+    ...mapActions('appbar', ['setTitle', 'setTabs', 'setShowBackButton']),
     ...mapActions('progressindicator', ['setLoading'])
   }
 })
@@ -108,6 +108,7 @@ export default class ConfigurationNewPage extends Vue {
   setTabs!: SetTabsAction
   setTitle!: SetTitleAction
   setLoading!: SetLoadingAction
+  setShowBackButton!: SetShowBackButtonAction
 
   created () {
     this.initializeAppBar()
@@ -139,6 +140,9 @@ export default class ConfigurationNewPage extends Vue {
   }
 
   initializeAppBar () {
+    if ('from' in this.$route.query && this.$route.query.from === 'searchResult') {
+      this.setShowBackButton(true)
+    }
     this.setTabs([
       {
         to: '/configurations/new',